upload progressbar implemented

+++ app/src/main/java/org/take365/Models/StoryDay.java
@@ -8,6 +8,6 @@ import org.take365.Engine.Network.Models.StoryImageImagesModel;
 
 public class StoryDay {
     public String day;
-    public float uploadProgress;
+    public int uploadProgress;
     public StoryImageImagesModel image;
 }
